python/pyproject.toml
[build-system]
requires = ["hatchling", "hatch.vcs"]
build-backend = "hatchling.build"

[project]
dynamic = ["version"]
name = "greenbids-tailor"
dependencies = [
    "fastapi~=0.111",
    "opentelemetry-api~=1.25",
    "opentelemetry-sdk~=1.25",
    "opentelemetry-exporter-otlp-proto-grpc~=1.25",
    "opentelemetry-instrumentation-fastapi~=0.46b0",
]
requires-python = ">=3.11"
authors = [{ name = "greenbids.ai" }]
description = "Bring traffic shaping to your own cloud"
readme = "../README.md"
license = { file = "../LICENSE" }
keywords = ["traffic shaping", "adtech", "openrtb"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
]

[project.urls]
Repository = "https://github.com/greenbids/greenbids-tailor-external.git"

[project.scripts]
greenbids-tailor = "greenbids.tailor.core.__main__:run"

[project.entry-points.greenbids-tailor-models]
None = 'greenbids.tailor.core.models'

[tool.hatch.build.targets.wheel]
packages = ["src/greenbids"]

[tool.hatch.build.hooks.vcs]
version-file = "src/greenbids/tailor/core/_version.py"

[tool.hatch.version]
source = "vcs"
[tool.hatch.version.raw-options]
root = ".."
